fix: UIの改善、一部TypeScriptのJSX化
