+++ front/my-app/app/dashboard/components/RandomWord.jsx
@@ -1,6 +1,6 @@
 import * as React from "react";
 
-const RandomWord: React.FC = () => {
+const RandomWord = () => {
   // 静的なランダムメッセージの配列
   // VSコードがstingと理解しているのでコロンStrongは不要。基本的に型推論がつかないときだけ型を明示するようにする。
   const randomWords = [
@@ -27,11 +27,11 @@ const RandomWord: React.FC = () => {
   ];
 
   // ランダムなインデックスを生成
-  const randomIndex: number = Math.floor(Math.random() * randomWords.length);
-  const word: string = randomWords[randomIndex];
+  const randomIndex = Math.floor(Math.random() * randomWords.length);
+  const word = randomWords[randomIndex];
 
   return (
-    <div className="mb-2 p-4 w-44 h-52 bg-white shadow-custom-dark rounded-3xl flex items-center justify-center text-center">
+    <div className="px-4 w-48 h-52 bg-white shadow-custom-dark rounded-3xl flex items-center justify-center text-center">
       <p className="font-bold">{word}</p>
     </div>
   );
