+++ front/my-app/app/dashboard/page.jsx
@@ -13,18 +13,13 @@ import ViewTable from "./components/ViewTable";
 const Dashboard = () => {
   return (
     <>
-      <div className="mx-7 grid 845s:grid-cols-2 1320s:grid-cols-3 1600s:grid-cols-4 gap-5 place-items-center mt-12">
+      <div className="mx-7 grid 845s:grid-cols-2 1600s:grid-cols-4 gap-5 place-items-center mt-12">
         <div>
-          <div className="flex space-x-2">
+          <div className="flex space-x-4">
             <MonthHWAT />
             <RandomWord />
           </div>
-          <div className="hidden 1320s:flex justify-center items-center bg-orange-200 rounded-3xl shadow-custom-dark">
-            <TsunaRhythm />
-          </div>
-        </div>
-        <div className="bg-orange-200 rounded-3xl shadow-custom-dark block 1320s:hidden">
-          <TsunaRhythm />
+          <ViewTable />
         </div>
         <div className="space-y-2">
           <GraphWorkingMinutes />
@@ -38,13 +33,10 @@ const Dashboard = () => {
           <ProjectHourlyWageRanking />
           <WorkTypeHourlyWageRanking />
         </div>
-        <Calendar />
-        <div className="col-span-2">
-          <ViewTable />
-        </div>
-      </div>
-      <div className="block 845s:hidden mt-4">
-        <TsunaRhythm />
+        {/* <Calendar /> */}
+        {/* <div className="col-span-2 justify-center items-center bg-orange-200 rounded-3xl shadow-custom-dark">
+          <TsunaRhythm />
+        </div> */}
       </div>
     </>
   );
